Extract run safely external commands
