+++ perun/collect/trace/systemtap/engine.py
@@ -29,6 +29,7 @@ import perun.utils as utils
 import perun.utils.helpers as helpers
 import perun.utils.log as perun_log
 import perun.utils.metrics as metrics
+from perun.utils.external import commands
 from perun.logic.locks import LockType, ResourceLock, get_active_locks_for
 from perun.utils.exceptions import (
     SystemTapStartupException,
@@ -193,7 +194,7 @@ class SystemTapEngine(engine.CollectEngine):
 
         # Run the compilation process
         # Fetch the password so that the preexec_fn doesn't halt
-        utils.run_safely_external_command("sudo sleep 0")
+        commands.run_safely_external_command("sudo sleep 0")
         # Run only the first 4 phases of the stap command, before actually running the collection
         with utils.nonblocking_subprocess(
             command + " -p 4",
@@ -382,7 +383,7 @@ class SystemTapEngine(engine.CollectEngine):
             # Form the module name which consists of the base module name and stapio PID
             module_name = f"{self.stap_module}__{self.stapio}"
             # Attempts to unload the module
-            utils.run_safely_external_command(f"sudo rmmod {module_name}", False)
+            commands.run_safely_external_command(f"sudo rmmod {module_name}", False)
             if not _wait_for_resource_release(_loaded_stap_kernel_modules, [module_name]):
                 WATCH_DOG.debug(f"Unloading the kernel module '{module_name}' failed")
         finally:
@@ -401,7 +402,7 @@ def _extract_usdt_probes(binary):
 
     :return str: the decoded standard output
     """
-    out, _ = utils.run_safely_external_command(
+    out, _ = commands.run_safely_external_command(
         f'sudo stap -l \'process("{binary}").mark("*")\'', False
     )
     return out.decode("utf-8")
@@ -568,7 +569,7 @@ def _extract_processes(extract_command):
                   attributes of the extracted processes
     """
     procs = []
-    out = utils.run_safely_external_command(extract_command, False)[0].decode("utf-8")
+    out = commands.run_safely_external_command(extract_command, False)[0].decode("utf-8")
     for line in out.splitlines():
         process_record = line.split()
 
@@ -604,7 +605,7 @@ def _loaded_stap_kernel_modules(module=None):
     extractor = f"lsmod | grep {module_filter} | awk '{{print $1}}'"
 
     # Run the command and save the found modules
-    out, _ = utils.run_safely_external_command(extractor, False)
+    out, _ = commands.run_safely_external_command(extractor, False)
     # Make sure that we have a list of unique modules
     modules = set()
     for line in out.decode("utf-8").splitlines():
