+++ tests/test_fuzzing.py
@@ -9,7 +9,6 @@ import pytest
 
 from click.testing import CliRunner
 
-import perun.utils as utils
 import perun.cli as cli
 import perun.fuzz.evaluate.by_coverage as coverage_fuzz
 import perun.fuzz.evaluate.by_perun as perun_fuzz
@@ -41,7 +40,7 @@ def test_fuzzing_coverage(capsys):
     command = " ".join([os.path.abspath(hang_test), num_workload])
     out, _ = capsys.readouterr()
 
-    utils.run_safely_external_command(command)
+    commands.run_safely_external_command(command)
     cov = coverage_fuzz.get_coverage_from_dir(os.getcwd(), coverage_config)
     assert cov != 0
 
@@ -531,7 +530,7 @@ def test_fuzzing_errors(pcs_with_root, monkeypatch):
     tail = os.path.join(examples, "tail", "tail")
 
     # Test when target testing returns error
-    old_run_process = utils.run_safely_external_command
+    old_run_process = commands.run_safely_external_command
 
     def patched_run_process(*_, **__):
         caller = sys._getframe().f_back.f_code.co_name
@@ -545,7 +544,7 @@ def test_fuzzing_errors(pcs_with_root, monkeypatch):
     def patched_check_output(*_, **__):
         return "real 0.01\nuser 0.00\nsys 0.00"
 
-    monkeypatch.setattr(utils, "run_safely_external_command", patched_run_process)
+    monkeypatch.setattr(commands, "run_safely_external_command", patched_run_process)
     monkeypatch.setattr(commands, "get_stdout_from_external_command", patched_check_output)
     result = runner.invoke(
         cli.fuzz_cmd,
@@ -566,7 +565,7 @@ def test_fuzzing_errors(pcs_with_root, monkeypatch):
     )  # fmt: skip
     asserts.predicate_from_cli(result, result.exit_code == 0)
     asserts.predicate_from_cli(result, "Faults: 0" not in result.output)
-    monkeypatch.setattr(utils, "run_safely_external_command", old_run_process)
+    monkeypatch.setattr(commands, "run_safely_external_command", old_run_process)
 
     # Test when target testing returns error
     old_target_perun_testing = perun_fuzz.target_testing
